requirements.txt
PyQt6>=6.7
numpy>=1.26
Pillow>=10.0
sounddevice>=0.4
